requirement.txt
absl-py==0.8.0
apturl==0.5.2
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==3.1.0
blinker==1.3
Brlapi==0.6.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
entrypoints==0.3
feedparser==5.1.3
future==0.18.1
gast==0.3.2
gluoncv==0.5.0
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
graphviz==0.8.4
grpcio==1.24.1
guacamole==0.9.2
h5py==2.10.0
html5lib==0.999
httplib2==0.9.1
hypothesis==4.42.0
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
ipdb==0.12.2
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.8
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-resnet==0.1.0
keras-retinanet==0.5.1
kiwisolver==1.1.0
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.1.1
MarkupSafe==0.23
matplotlib==3.0.3
mistune==0.8.4
more-itertools==7.2.0
mxnet==1.5.1.post0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.17.2
oauthlib==3.1.0
onboard==1.2.0
opencv-python==4.1.1.26
padme==1.1.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.0.1
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
plainbox==0.25
plotly==4.2.1
progressbar2==3.47.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.2.7
pycups==1.9.73
pycurl==7.43.0
Pygments==2.4.2
pygobject==3.20.0
PyJWT==1.3.0
pykalman==0.9.5
pyparsing==2.0.3
pyrsistent==0.15.6
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.8.0
python-debian==0.1.27
python-mnist==0.6
python-systemd==231
python-utils==2.3.0
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
reportlab==3.3.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
six==1.10.0
sklearn==0.0
system-service==0.3
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.6
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zipp==0.6.0
